factor out stellar-halo-fitting function
